add hr navigation bar
